Seperated vulnerabilities patterns from main code.
All the vulnerability patterns are in patterns.json now, which the logscanner.js function will take in from.

This improves the scalability and mantainance of code, as you can just expand on the  cves covered by just adding to the patterns.json file without touching the logscanner.js main code.